Improve ManufacturerData
